+++ randovania/generator/item_pool/major_items.py
@@ -34,7 +34,7 @@ def add_major_items(resource_database: ResourceDatabase,
 
     item_pool: list[PickupEntry] = []
     new_assignment: dict[PickupIndex, PickupEntry] = {}
-    initial_resources = ResourceCollection.with_database(resource_database)
+    starting = []
 
     for item, state in major_items_configuration.items_state.items():
         if len(item.ammo_index) != len(state.included_ammo):
@@ -56,8 +56,6 @@ def add_major_items(resource_database: ResourceDatabase,
             item_pool.append(create_major_item(item, state, True, resource_database, ammo, locked_ammo))
 
         for _ in range(state.num_included_in_starting_items):
-            initial_resources.add_resource_gain(
-                create_major_item(item, state, False, resource_database, ammo, locked_ammo).all_resources,
-            )
+            starting.append(create_major_item(item, state, False, resource_database, ammo, locked_ammo))
 
-    return PoolResults(item_pool, new_assignment, initial_resources)
+    return PoolResults(item_pool, new_assignment, starting)