+++ randovania/layout/description_migration.py
@@ -270,11 +270,19 @@ def _migrate_v12(json_dict: dict) -> dict:
 
     return json_dict
 
+
 def _migrate_v13(json_dict: dict) -> dict:
     for game in json_dict["game_modifications"]:
         area_identifier = AreaIdentifier.from_string(game["starting_location"])
         node_identifier = migration_data.get_new_start_loc_from_old_start_loc(game["game"], area_identifier)
         game["starting_location"] = node_identifier.as_string
+
+    return json_dict
+
+
+def _migrate_v14(json_dict: dict) -> dict:
+    for game in json_dict["game_modifications"]:
+        game["starting_equipment"] = {"items": game.pop("starting_items")}
 
     return json_dict
 
@@ -293,6 +301,7 @@ _MIGRATIONS = [
     _migrate_v11,
     _migrate_v12,
     _migrate_v13,
+    _migrate_v14,
 ]
 CURRENT_VERSION = migration_lib.get_version(_MIGRATIONS)